fix(polyfill): Do not monkey patch Symbol and WeakMap polyfills onto global object
