+++ src/utils/Symbol.js
@@ -1,0 +1,10 @@
+// https://github.com/Riim/symbol-polyfill/blob/master/index.js
+
+export default typeof Symbol === 'function' ? Symbol : (() => {
+  let idCounter = 0
+  function Symbol (key) {
+    return '__' + key + '_' + Math.floor(Math.random() * 1e9) + '_' + (++idCounter) + '__'
+  }
+  Symbol.iterator = Symbol('Symbol.iterator')
+  return Symbol
+})()
